NAT: Support multiple tenants with overlapping address space (multiple context ids) share a single NAT pool.
Change the API to tie a NAT pool to a particular context id.
If a tunnel is bound to the same context as another, if a packet with the same
5-tuple is sent on both tunnels, the same session is used so return traffic will
only go back on the first tunnel. No form of load balancing will happen.

If a tunnel belongs to a different context (via a different tenant), two separate
sessions (and a separate NAT mapping) will be created.

Signed-off-by: Ole Troan <otroan@employees.org>