Add feature tests for TCP session reopening.
A TCP session is established as soon as we see an ACK in both directions.
A session is only reopened if we see a SYN from the inside.
Fix handling the SYN flag.

NB: The difference in new an reopened state machine should not lead to issues.
But time will show.

Signed-off-by: Ole Troan <otroan@employees.org>